+++ My.DAL/Data/Migrations/20260902140000_AddGoogleCalendarAvailabilityOnly.cs
@@ -6,6 +6,9 @@ using My.DAL.Data;
 
 namespace My.DAL.Data.Migrations
 {
+    /// <summary>
+    /// Already on UserSettings in InitialMigration for greenfield; conditional add for upgrades.
+    /// </summary>
     [DbContext(typeof(ApplicationDbContext))]
     [Migration("20260902140000_AddGoogleCalendarAvailabilityOnly")]
     public partial class AddGoogleCalendarAvailabilityOnly : Migration
@@ -14,11 +17,15 @@ namespace My.DAL.Data.Migrations
         protected override void Up(MigrationBuilder migrationBuilder)
         {
             migrationBuilder.Sql("""
-                ALTER TABLE [UserSettings] ADD [GoogleCalendarAvailabilityOnly] bit NOT NULL
-                    CONSTRAINT [DF_UserSettings_GoogleCalendarAvailabilityOnly] DEFAULT 0;
+                IF COL_LENGTH('UserSettings', 'GoogleCalendarAvailabilityOnly') IS NULL
+                BEGIN
+                    ALTER TABLE [UserSettings] ADD [GoogleCalendarAvailabilityOnly] bit NOT NULL
+                        CONSTRAINT [DF_UserSettings_GoogleCalendarAvailabilityOnly] DEFAULT 0;
+                END
                 """);
             migrationBuilder.Sql("""
-                ALTER TABLE [UserSettings] DROP CONSTRAINT [DF_UserSettings_GoogleCalendarAvailabilityOnly];
+                IF OBJECT_ID(N'[DF_UserSettings_GoogleCalendarAvailabilityOnly]', N'D') IS NOT NULL
+                    ALTER TABLE [UserSettings] DROP CONSTRAINT [DF_UserSettings_GoogleCalendarAvailabilityOnly];
                 """);
         }
 
@@ -26,7 +33,8 @@ namespace My.DAL.Data.Migrations
         protected override void Down(MigrationBuilder migrationBuilder)
         {
             migrationBuilder.Sql("""
-                ALTER TABLE [UserSettings] DROP COLUMN [GoogleCalendarAvailabilityOnly];
+                IF COL_LENGTH('UserSettings', 'GoogleCalendarAvailabilityOnly') IS NOT NULL
+                    ALTER TABLE [UserSettings] DROP COLUMN [GoogleCalendarAvailabilityOnly];
                 """);
         }
     }
